+++ src/features/XIT/ACT/action-steps/CONT_SEND.ts
@@ -1,19 +1,21 @@
 import { act } from '@src/features/XIT/ACT/act-registry';
 import { fixed0 } from '@src/utils/format';
-import {
-  changeInputValue,
-  changeSelectIndex,
-  changeTextAreaValue,
-  clickElement,
-  focusElement,
-} from '@src/util';
-import { contractDraftsStore } from '@src/infrastructure/prun-api/data/contract-drafts';
+import { changeInputValue, changeSelectIndex, focusElement } from '@src/util';
 import { materialsStore } from '@src/infrastructure/prun-api/data/materials';
-import { $ } from '@src/utils/select-dom';
 import {
   waitFor,
   selectLocation,
-  selectMaterial,
+  createNewDraft,
+  setDraftNameAndPreamble,
+  saveDraftDetails,
+  openTemplate,
+  selectTemplateType,
+  setCurrency,
+  addMaterials,
+  setDeadline,
+  applyTemplate,
+  saveConditions,
+  ContDraftContext,
 } from '@src/features/XIT/ACT/action-steps/cont-utils';
 
 interface Data {
@@ -41,41 +43,10 @@ export const CONT_SEND = act.addActionStep<Data>({
     const listTile = await requestTile('CONTD');
     if (!listTile) return;
 
-    setStatus('Creating new contract draft...');
+    const draftCtx: ContDraftContext = { draftTile: listTile, log, setStatus, fail };
 
-    const isCreateNew = (btn: Element) => btn.textContent?.trim().toLowerCase() === 'create new';
-
-    const findContdButton = () => {
-      for (const tile of tiles.find('CONTD', true)) {
-        const btn = _$$(tile.anchor, C.Button.btn).find(isCreateNew);
-        if (btn) return { tile, btn };
-      }
-      return undefined;
-    };
-
-    const createBtnReady = await waitFor(() => !!findContdButton(), 10000);
-    if (!createBtnReady) {
-      fail('Could not find "Create New" button');
-      return;
-    }
-
-    const { btn: createBtn } = findContdButton()!;
-
-    const beforeIds = new Set((contractDraftsStore.all.value ?? []).map(d => d.naturalId));
-    await clickElement(createBtn);
-
-    setStatus('Waiting for draft to be created...');
-    const draftAppeared = await waitFor(
-      () => (contractDraftsStore.all.value ?? []).some(d => !beforeIds.has(d.naturalId)),
-      8000,
-    );
-    if (!draftAppeared) {
-      fail('Timed out waiting for new contract draft');
-      return;
-    }
-
-    const newDraft = (contractDraftsStore.all.value ?? []).find(d => !beforeIds.has(d.naturalId))!;
-    log.info(`New draft created: ${newDraft.naturalId}`);
+    const newDraft = await createNewDraft(draftCtx);
+    if (!newDraft) return;
 
     setStatus(`Loading draft ${newDraft.naturalId}...`);
     const draftTile = await requestTile(`CONTD ${newDraft.naturalId}`);
@@ -105,21 +76,12 @@ export const CONT_SEND = act.addActionStep<Data>({
       );
     }
 
-    setStatus('Setting contract name...');
+    // Update context to point at the actual draft tile.
+    const ctx2: ContDraftContext = { draftTile, log, setStatus, fail };
 
     const now = new Date();
     const dateStr = now.toLocaleDateString('en-US', { month: 'short', day: 'numeric' });
     const contractName = `${data.packageName} - ${data.contDest ?? ''} - ${dateStr}`;
-
-    const nameInput = (await $(draftTile.anchor, 'input')) as HTMLInputElement | null;
-    if (nameInput) {
-      focusElement(nameInput);
-      nameInput.select();
-      changeInputValue(nameInput, contractName);
-      log.info(`Name set: ${contractName}`);
-    } else {
-      log.warning('Could not find name input');
-    }
 
     const materialsList = materialDetails.map(m => `${m.ticker} x${m.amount}`).join(', ');
     const preambleText =
@@ -131,123 +93,19 @@ export const CONT_SEND = act.addActionStep<Data>({
           : '') +
         (data.daysToFulfill > 0 ? `Delivery within ${data.daysToFulfill} days` : '');
 
-    const preambleInput = _$(draftTile.anchor, 'textarea') as HTMLTextAreaElement | null;
-    if (preambleInput) {
-      focusElement(preambleInput);
-      changeTextAreaValue(preambleInput, preambleText);
-      log.info('Preamble set');
-    }
+    await setDraftNameAndPreamble(ctx2, contractName, preambleText);
+    await saveDraftDetails(ctx2);
 
-    setStatus('Saving draft details...');
+    const templateSelect = await openTemplate(ctx2);
+    if (!templateSelect) return;
 
-    const saveBtn = _$$(draftTile.anchor, C.Button.btn).find(
-      (btn: HTMLElement) => btn.textContent?.trim().toLowerCase() === 'save',
-    ) as HTMLElement | undefined;
-    if (saveBtn) {
-      await clickElement(saveBtn);
-      log.info('Draft details saved');
-    } else {
-      log.warning('Could not find save button for draft details');
-    }
+    selectTemplateType(ctx2, templateSelect, 'SHIP');
+    await setCurrency(ctx2, data.currency);
 
-    setStatus('Opening template selection...');
-
-    const selectTemplateBtnReady = await waitFor(
-      () =>
-        _$$(draftTile.anchor, 'button').some(
-          btn => btn.textContent?.trim().toLowerCase() === 'select template',
-        ),
-      5000,
+    await addMaterials(
+      ctx2,
+      materialDetails.map(m => ({ ticker: m.ticker, amount: m.amount })),
     );
-    if (!selectTemplateBtnReady) {
-      fail('Could not find "Select Template" button');
-      return;
-    }
-    const selectTemplateBtn = _$$(draftTile.anchor, 'button').find(
-      btn => btn.textContent?.trim().toLowerCase() === 'select template',
-    )!;
-    await clickElement(selectTemplateBtn);
-
-    setStatus('Selecting Ship commodity template...');
-
-    const templateTypeContainer = await $(draftTile.anchor, C.TemplateSelection.templateTypeSelect);
-    const templateSelect = _$(templateTypeContainer, 'select') as HTMLSelectElement | null;
-    if (!templateSelect) {
-      fail('Could not find template type select');
-      return;
-    }
-
-    const shipIndex = Array.from(templateSelect.options).findIndex(o => o.value === 'SHIP');
-    if (shipIndex >= 0) {
-      changeSelectIndex(templateSelect, shipIndex);
-    }
-    log.info('Selected "Ship commodity" template');
-
-    const currencySelectFound = await waitFor(() => {
-      const selects = _$$(draftTile.anchor, 'select') as HTMLSelectElement[];
-      return selects.some(s => Array.from(s.options).some(o => o.value === data.currency));
-    }, 3000);
-
-    if (currencySelectFound) {
-      const selects = _$$(draftTile.anchor, 'select') as HTMLSelectElement[];
-      const currencySelect = selects.find(s =>
-        Array.from(s.options).some(o => o.value === data.currency),
-      )!;
-      const currencyIndex = Array.from(currencySelect.options).findIndex(
-        o => o.value === data.currency,
-      );
-      if (currencyIndex >= 0) {
-        changeSelectIndex(currencySelect, currencyIndex);
-      }
-      log.info(`Currency set to ${data.currency}`);
-    } else {
-      log.warning(`Could not find currency select for ${data.currency}`);
-    }
-
-    setStatus('Adding materials to template...');
-
-    for (let i = 0; i < materialDetails.length; i++) {
-      const mat = materialDetails[i];
-
-      if (i > 0) {
-        const addBtn = _$$(draftTile.anchor, 'button').find(btn => {
-          const t = btn.textContent?.trim().toLowerCase();
-          return t === 'add shipment' || t === 'add commodity';
-        });
-        if (!addBtn) {
-          log.warning(`Could not find add button for ${mat.ticker}`);
-          continue;
-        }
-        await clickElement(addBtn);
-        await waitFor(() => _$$(draftTile.anchor, C.TemplateSelection.group).length >= i + 1, 2000);
-      }
-
-      const groups = _$$(draftTile.anchor, C.TemplateSelection.group);
-      const group = groups.at(-1);
-      if (!group) {
-        log.warning(`Could not find group for ${mat.ticker}`);
-        continue;
-      }
-
-      const amountInput = group.querySelector(
-        'input[inputmode="numeric"]',
-      ) as HTMLInputElement | null;
-      if (amountInput) {
-        focusElement(amountInput);
-        amountInput.select();
-        changeInputValue(amountInput, String(mat.amount));
-      }
-
-      const matSelectorContainer = _$(group, C.MaterialSelector.container);
-      if (matSelectorContainer) {
-        const ok = await selectMaterial(matSelectorContainer, mat.ticker);
-        if (ok) {
-          log.info(`Added: ${mat.ticker} x${mat.amount}`);
-        } else {
-          log.warning(`Could not select material ${mat.ticker}`);
-        }
-      }
-    }
 
     // The SHIP template price field is per-commodity — the game charges this
     // amount for each commodity row. Divide total payment by number of commodities.
@@ -310,53 +168,12 @@ export const CONT_SEND = act.addActionStep<Data>({
       }
     }
 
-    if (data.daysToFulfill > 0) {
-      const deadlineInput = draftTile.anchor.querySelector(
-        'input[name="deadline"]',
-      ) as HTMLInputElement | null;
-      if (deadlineInput) {
-        focusElement(deadlineInput);
-        deadlineInput.select();
-        changeInputValue(deadlineInput, String(data.daysToFulfill));
-        log.info(`Deadline set: ${data.daysToFulfill} days`);
-      }
-    }
+    setDeadline(ctx2, data.daysToFulfill);
 
-    setStatus('Applying template...');
+    const applied = await applyTemplate(ctx2);
+    if (!applied) return;
 
-    const applyBtnReady = await waitFor(
-      () =>
-        _$$(draftTile.anchor, 'button').some(
-          btn => btn.textContent?.trim().toLowerCase() === 'apply template',
-        ),
-      5000,
-    );
-    if (!applyBtnReady) {
-      fail('Could not find "Apply Template" button');
-      return;
-    }
-    const applyBtn = _$$(draftTile.anchor, 'button').find(
-      btn => btn.textContent?.trim().toLowerCase() === 'apply template',
-    )!;
-
-    await clickElement(applyBtn);
-    await waitFor(() => applyBtn.classList.contains(C.Button.disabled), 3000);
-    await waitFor(() => !applyBtn.classList.contains(C.Button.disabled), 5000);
-    log.info('Template applied');
-
-    // Pause so the user can review the filled contract before it is saved.
-    await waitAct('Save conditions?');
-    setStatus('Saving conditions...');
-
-    const condSaveBtn = _$$(draftTile.anchor, C.Button.btn).findLast(
-      (btn: HTMLElement) => btn.textContent?.trim().toLowerCase() === 'save',
-    ) as HTMLElement | undefined;
-    if (condSaveBtn && !condSaveBtn.classList.contains(C.Button.disabled)) {
-      await clickElement(condSaveBtn);
-      log.info('Conditions saved');
-    } else {
-      log.warning('Conditions save button not found or disabled');
-    }
+    await saveConditions(ctx2, waitAct);
 
     log.success(`Contract draft ${newDraft.naturalId} ready to send`);
     complete();
