+++ src/features/XIT/ACT/action-steps/MTRA_TRANSFER.ts
@@ -1,11 +1,12 @@
 import { act } from '@src/features/XIT/ACT/act-registry';
 import { serializeStorage } from '@src/features/XIT/ACT/actions/utils';
 import { fixed0 } from '@src/utils/format';
-import { changeInputValue, clickElement, focusElement } from '@src/util';
+import { changeInputValue, clickElement } from '@src/util';
 import { materialsStore } from '@src/infrastructure/prun-api/data/materials';
 import { watchWhile } from '@src/utils/watch';
 import { storagesStore } from '@src/infrastructure/prun-api/data/storage';
 import { AssertFn } from '@src/features/XIT/ACT/shared-types';
+import { selectMaterial } from '@src/features/XIT/ACT/action-steps/cont-utils';
 
 interface Data {
   from: string;
@@ -68,26 +69,12 @@ export const MTRA_TRANSFER = act.addActionStep<Data>({
 
     setStatus('Setting up MTRA buffer...');
     const container = await $(tile.anchor, C.MaterialSelector.container);
-    const input = await $(container, 'input');
 
-    const suggestionsContainer = await $(container, C.MaterialSelector.suggestionsContainer);
-    focusElement(input);
-    changeInputValue(input, ticker);
-
-    const suggestionsList = await $(container, C.MaterialSelector.suggestionsList);
-    suggestionsContainer.style.display = 'none';
-    const match = _$$(suggestionsList, C.MaterialSelector.suggestionEntry).find(
-      x => _$(x, C.ColoredIcon.label)?.textContent === ticker,
-    );
-
-    if (!match) {
-      suggestionsContainer.style.display = '';
+    const ok = await selectMaterial(container, ticker);
+    if (!ok) {
       fail(`Ticker ${ticker} not found in the material selector`);
       return;
     }
-
-    await clickElement(match);
-    suggestionsContainer.style.display = '';
 
     const sliderNumbers = _$$(tile.anchor, 'rc-slider-mark-text').map(x =>
       Number(x.textContent ?? 0),
